Aggressive commit
Finally changes 1024 to 1000. I just needed some time for it
Paid should generally be Double to avoid overflow/parse errors from JSON
Bad patch to manage anorak

+++ app/src/main/java/it/angelic/mpw/MinersActivity.java
@@ -185,7 +185,7 @@ public class MinersActivity extends DrawerActivity {
     @NonNull
     private ArrayList<MinerDBRecord> updateUIRecordStats(ArrayList<MinerDBRecord> minerDbList) {
         long hihr = 0;
-        Long hiPaid = 0l;
+        Double hiPaid = 0d;
         Integer hiPaidIdx = null;
         Date oldestDt = new Date();
         Integer minDateSeeIdx = null;
@@ -198,7 +198,7 @@ public class MinersActivity extends DrawerActivity {
                 hihrIdx = cnt;
             }
             //DB only
-            long minp = minerDbrec.getPaid() == null ? 0 : minerDbrec.getPaid();
+            double minp = minerDbrec.getPaid() == null ? 0 : minerDbrec.getPaid();
             if (minp > hiPaid) {
                 hiPaid = minp;
                 hiPaidIdx = cnt;
